requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
SQLAlchemy==2.0.36

# Vector DB
chromadb==0.5.20

# Data processing
pandas==2.2.3
python-dateutil==2.9.0

# File handling
python-multipart==0.0.9
aiofiles==23.2.1

# HTTP client
httpx==0.27.2

# Config
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
